+++ CHANGES.rst
@@ -5,7 +5,12 @@ Changelog
 0.8.dev0
 --------
 
-- no changes yet.
+- Add ajax overlay with large product image in product view when clicking on
+  thumb.
+  [rnix]
+
+- Improve tiles view.
+  [rnix]
 
 
 0.7